finished state of Project before abandoment

+++ index.js
@@ -1,22 +1,26 @@
 import express from "express"
 import fs from "fs"
 import path from "path"
-import sharp from "sharp" // sharp is a library for image processing
 import multer from "multer"
 import ImageClassifier from "./ai/ImageClassifier.js"
 
 const aiImageClassifier = new ImageClassifier()
-const aiModelPath = path.join("ai", "model.json")
+
+//setting this to false is pretty good to generate Data by Drawing
+const doTraining = true // Enables/Disables Training and File deletion
 
 console.log("training AI locally")
-await aiImageClassifier.train(path.join("ai", "test"))
+if (doTraining)
+  await aiImageClassifier.train(
+    path.join("ai", "datasets"),
+    path.join("ai", "test")
+  )
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
-    cb(null, "./uploads/") // Specify the destination folder
+    cb(null, "./uploads/") // Save the file to the uploads directory
   },
   filename: function (req, file, cb) {
-    // TODO Generate a unique filename for each uploaded file
     cb(null, `${Date.now()}-${path.basename(file.originalname)}`)
   },
 })
@@ -38,15 +42,14 @@ app.post(
       console.log("File received")
       const file = req.files.image[0]
       const filePath = file.path
-      // Assuming you have a function sendToAI that sends the image to your AI service and returns a promise
-      const aiResult = await sendToAI(filePath)
 
-      // Delete the original and grayscale images after processing
-      fs.unlinkSync(filePath)
+      const aiResult = doTraining ? await sendToAI(filePath) : {}
+
+      // Delete the Image after processing
+      if (doTraining) fs.unlinkSync(filePath)
 
       // Send the AI analysis result back to the client
-      res.json({ message: "Image uploaded successfully", aiResult })
-      // res.json({ message: "succ" })
+      res.json({ message: "Image uploaded successfully", ...aiResult })
     } catch (error) {
       console.error(`Error processing image: ${error}`)
       res.status(500).send("Failed to process the image")
@@ -58,10 +61,8 @@ app.listen(3000, () =>
   console.log("Server running on port http://127.0.0.1:3000/")
 )
 
-//dataset
-//schawrz-weiss /// pixel oder grid /// auswerten
 async function sendToAI(imagePath) {
-  const res = await aiImageClassifier.predict(imagePath)
-  console.log("res:", res)
-  return { successful: true, data: res, test: Math.floor(Math.random() * 10) }
+  const isRaw = true
+  const res = await aiImageClassifier.predict(imagePath, isRaw)
+  return { successful: true, data: res, isRaw: isRaw }
 }
